change EventManager to Even class
