+++ DiceRoll/CombatEvent.cs
@@ -1,57 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
+
+using System;
 
 namespace Monobius
 {
-    public class EventManager
+    public class CombatEvent : Event
     {
-        //ALL EVENT TYPES//
-        public enum Events
-        {
-            Combat,
-            Treasure,
-            Lose,
-            Win
-        }
-
-        public Events Type;
-        public bool IsDecrypted = false;
-        public bool IsEventConcluded = false;
         public int RoundCounter = 0;
 
-        public EventManager (Events type)
-        {
-            Type = type;
-        }
-        public void Execute(GameManagerV2 gm)
-        {
-            if (IsDecrypted)
-            {
-                gm.Dialog.FailDecryption();
-                return;
-            }
-
-            switch (gm.CurrentPlayerRoom.Event.Type)
-            {
-                case EventManager.Events.Combat:
-                    Combat(gm);
-                    break;
-                case EventManager.Events.Treasure:
-                    Treasure(gm);
-                    break;
-                case EventManager.Events.Lose: 
-                    Lose(gm); 
-                    break;
-                case EventManager.Events.Win:
-                    Win(gm);
-                    break;
-                default:
-                    throw new NotImplementedException();
-            }
-
-            IsDecrypted = true;
-        }
-        public void Combat(GameManagerV2 gm)
+        public override void Execute(GameManagerV2 gm)
         {
             gm.Player.Vessel.ResetWeapons();
             Vessel enemyVessel = gm.Map.EnemyVessels[0];
@@ -76,12 +32,13 @@ namespace Monobius
                 Win(gm);
             }
         }
+
         public void CombatRound(GameManagerV2 gm, Vessel enemyVessel)
         {
 
             int playerVesselWeaponIndex = -1;
 
-            while (playerVesselWeaponIndex <0)
+            while (playerVesselWeaponIndex < 0)
             {
                 gm.Dialog.Write("Current ASSETS Online: ");
                 gm.Dialog.Write("----------------------");
@@ -139,11 +96,12 @@ namespace Monobius
                     gm.Player.Vessel.Health -= dmg;
                     gm.Dialog.Write("");
                     gm.Dialog.Write("The " + gm.Player.Vessel.Name + " takes " + dmg + " damage.");
-                } else
+                }
+                else
                 {
                     gm.Dialog.Write("Your shields absord the weak attack.");
                 }
-                
+
 
             }
             //IF BOTH SHIPS ROLL THE SAME//
@@ -172,104 +130,17 @@ namespace Monobius
                 Console.ReadLine();
             }
         }
-       
-        public void Treasure(GameManagerV2 gm)
-        {
-            int roomSelection = gm.Dice.Roll(3);
-            List<Weapon> weaponPool = gm.TreasurePoolWeapons;
-            List<Item> consumablePool = gm.TreasurePoolConsumable;
-            List<Item> passivePool = gm.TreasurePoolPassive;
-            Item selectedItem;
-            switch (roomSelection) 
-            {
-                case 1:
-                    {
-                        int poolSize = weaponPool.Count;
-                        int selectedIndex = gm.Dice.Roll(poolSize) - 1;
-                        Weapon weapon = weaponPool[selectedIndex];
-                        weaponPool.RemoveAt(selectedIndex);
-                        TreasureWeapon(gm, weapon);
-                    }
-                    break;
-                case 2:
-                    {
-                        int poolSize = consumablePool.Count;
-                        int selectedIndex = gm.Dice.Roll(poolSize) - 1;
-                        selectedItem = consumablePool[selectedIndex];
-                        consumablePool.RemoveAt(selectedIndex);
-                        TreasureItem(gm, selectedItem);
-                    }
-                    break;
-                case 3:
-                    {
-                        int poolSize = passivePool.Count;
-                        int selectedIndex = gm.Dice.Roll(poolSize) - 1;
-                        selectedItem = passivePool[selectedIndex];
-                        passivePool.RemoveAt(selectedIndex);
-                        TreasureItem(gm, selectedItem);
-                        selectedItem.Consume(gm);
-                    }
-                    break;
-                default:
-                    throw new NotImplementedException();
-            }
-
-
-        }
-        public void TreasureWeapon(GameManagerV2 gm, Weapon weapon)
-        {
-            bool isInvalidInput;
-            do
-            {
-                gm.Dialog.TreasureWeaponSelect0(weapon.Name);
-                isInvalidInput = false;
-                switch (gm.Dialog.Read())
-                {
-                    case "REPLACE":
-                    case "R":
-                        gm.Dialog.TreasureWeaponSelect1(gm);
-
-                        string userInput = gm.Dialog.Read();
-                        int weaponIndex = gm.Player.Vessel.GetWeaponIndexByName(userInput);
-                        if (weaponIndex < 0)
-                        {
-                            isInvalidInput = true;
-                        } else
-                        {
-                            gm.Player.Vessel.SetWeapon(weapon, weaponIndex);
-                        }
-                        break;
-                    case "DESTROY":
-                    case "D":
-                        gm.Dialog.TreasureWeaponSelect2();
-                        break;
-                    default:
-                        gm.Dialog.SelectionError();
-                        isInvalidInput = true;
-                        break;
-                }
-            } while (isInvalidInput);
-            
-
-            
-        }
-        public void TreasureItem(GameManagerV2 gm, Item item)
-        {
-            gm.Dialog.ItemSelect(item);
-            gm.Player.Inventory.AddItem(item);
-        }
 
         public void Lose(GameManagerV2 gm)
         {
-            //GAME OVER IF PLAYER DIES//
             gm.IsGameRunning = false;
             Console.ReadLine();
         }
+
         public void Win(GameManagerV2 gm)
         {
             gm.IsGameRunning = false;
-                Console.ReadLine();
+            Console.ReadLine();
         }
     }
 }
-
